+++ Runtime/NodeExtensions_Getter.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using AceLand.NodeFramework.Core;
@@ -6,30 +6,36 @@ using AceLand.TaskUtils;
 
 namespace AceLand.NodeFramework
 {
-    public partial class Node<T>
+    public static partial class NodeExtensions
     {
-        public static Task<T> GetAsync() =>
-            GetNode();
+        public static Task<T> GetAsync<T>(this INode<T> node) where T : class, INode =>
+            GetNode<T>();
+        
+        public static Task<T> GetAsync<T>(this INode<T> node, string id) where T : class, INode =>
+            GetNode<T>(id);
+        
+        public static Task<T> GetAsync<T, TEnum>(this INode<T> node, TEnum id) 
+            where T : class, INode 
+            where TEnum : Enum =>
+            GetNode<T>(id.ToString());
 
-        public static Task<T> GetAsync(string id) =>
-            GetNode(id);
-
-        public static Task<T> GetAsync<TEnum>(TEnum id) where TEnum : Enum =>
-            GetNode(id.ToString());
-
-        public static T Get() =>
-            Nodes.TryGetNode(out T node) ? node : null;
-
-        public static T Get(string id) =>
-            Nodes.TryGetNode(id, out T node) ? node : null;
-
-        public static T Get<TEnum>(TEnum id) where TEnum : Enum =>
-            Get(id.ToString());
-
-        public static IEnumerable<T> GetNodes() =>
+        public static T Get<T>(this INode<T> node) where T : class, INode =>
+            Nodes.TryGetNode(out T n) ? n : null;
+        
+        public static T Get<T>(this INode<T> node, string id) where T : class, INode =>
+            Nodes.TryGetNode(id, out T n) ? n : null;
+        
+        public static T Get<T, TEnum>(this INode<T> node, TEnum id)
+            where T : class, INode 
+            where TEnum : Enum =>
+            Get(node, id.ToString());
+        
+        public static IEnumerable<T> GetNodes<T>(this INode<T> node) where T : class, INode =>
             Nodes.GetNodesByType<T>();
-
-        private static async Task<T> GetNode()
+        
+        
+        // local functions
+        private static async Task<T> GetNode<T>() where T : class, INode
         {
             var aliveToken = Promise.ApplicationAliveToken;
             var targetTime = DateTime.Now.AddSeconds(NodeUtils.Settings.NodeGetterTimeout);
@@ -46,7 +52,7 @@ namespace AceLand.NodeFramework
             throw new Exception(msg);
         }
 
-        private static async Task<T> GetNode(string id)
+        private static async Task<T> GetNode<T>(string id) where T : class, INode
         {
             var aliveToken = Promise.ApplicationAliveToken;
             var targetTime = DateTime.Now.AddSeconds(NodeUtils.Settings.NodeGetterTimeout);