+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.0.8] - 2024-11-26
+
+### Modify
+- Getter To Extensions
+- Register To Extensions
+
 ## [1.0.7] - 2024-11-25
 
 ### Fixed